Avoid EPROBE_DEFER loop with Rockchip DSI driver
Previously, the Rockchip DSI driver would register itself as MIPI host
during probe, then proceed to register all sub-components, then
realize the panel isn't available yet, and defer.
This doesn't work with 6.1 though, and results in an endless loop.
The correct way is to let the DSI driver register as host, then wait for
the panel driver to be registered and probed, and then let the panel
driver trigger registering the DSI driver's sub-components.